+++ src/main/java/com/example/pdp_project/repo/MessageRepository.java
@@ -9,6 +9,6 @@ import java.util.List;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
     @Query("SELECT m FROM Message m WHERE (m.from.id = :user1 AND m.to.id = :user2) OR (m.from.id = :user2 AND m.to.id = :user1) ORDER BY m.sentAt ASC")
-    List<Message> getChatMessages(@Param("user1") Integer user1, @Param("user2") Integer user2);
+    List<Message> getChatMessages(@Param("user1") Long user1, @Param("user2") Long user2);
 
 }
